Disable Git clone inside a Git repo
In an effort to prevent users from accidentally cloning Git repos, the Git clone button is disabled if we detect that we are already inside a Git repo.

This is done using the onPathChanged() event from the file browser and using the existing `/git/all_history` API which is used to check for a Git repo in the main Git panel.

Also, add error handling when the git clone call fails to show the error message from the server.