Fix git repo determination when there are no tags
When there are no tags, `git describe` returns `fatal: No names found, cannot describe anything`. This fails the entire `/git/all_history` API call and we fail to identify the Git repo completely.

This commit doesn't fail in this case but returns an empty list of tags and also uses the `--tag` option which returns non-annotated tags as well.