Use data property to determine value of theme based icons
Instead of detecting if we are in a light theme via JS we can do this
with CSS by using the `data-them-light` attribute that is set by
JupyterLab.

This fixes a bug where if you changed the theme, the icons wouldn't
update until reloading, because the react components wouldn't know to
re-render on a theme change.

This ends up removing a fair amount of code as well.