Use the full path when opening the terminal from the Git menu
**Description**

Currently, the menu item opens based on the path of the current working directory relative to the server root. In some cases, the terminal starts from a directory other than the server root so this "cd $PATH" fails.

To avoid this, just construct the full path and "cd $SERVER_ROOT/$PATH" while executing the command

Other changes
* Moved the addMenuItems code into the gitMenuCommands file
* Removed unused command.