Server API to list changed files in Git
This change adds a endpoint to the server to list the files changed in the working tree, staging area, between two commits, or in a single commit.

This will be used by the "DiffComponent" to call the sub-components based on the file extension, i.e, all ipynb files will be delegated to the NBDiffComponent and all other files will be delegated to the PlainTextDiffComponent.