first pass at merging UI
creating backend for merge

Update disabled button

Make the text more informative + use transparency to indicate button as disabled.

minor styling improvements

made the mergeDialog box smaller + added a top border to the merge button in the panel so that it doesn't blend with the bottom branch when the bottom branch is selected.