don't prematurely close new branch dialog
if a name wasn't entered or an invalid branch name was entered then the
new branch dialog would close once you hit new branch. This makes it
so the create branch button is disabled when no name has been entered,
and keeps the dialog open until it is clear that branch creation was
successful.

I also removed the 'fatal' from the git error message as it was
redundant with the title of the dialog.