Allow the user to dismiss the modal
I am not convinced this is an entirely good idea, as why bother
with a blocking modal at all if a user can dismiss. The main hope
is that less technical users will simply respect the modal and wait
until completion before moving on. The risk is that users will
catch on and disregard the warning, allowing for potential footguns.
However, reviewers appear hostile to a blocking modal which cannot
be readily dismissed.